Conflicts:
	CMakeLists.txt
	Makefile.am
	springlobby.cbp
	src/chatpanel.cpp
	src/ibattle.cpp
	src/mainjoinbattletab.cpp
	src/mainoptionstab.cpp
	src/maintorrenttab.cpp
	src/playback/playbacktab.cpp
	src/widgets/downloadlistctrl.cpp

+++ src/battle.cpp
@@ -641,7 +641,7 @@ void Battle::SetInGame( bool value )
 	time_t now = time(0);
 	if ( m_ingame && !value )
 	{
-		for ( int i = 0; i < GetNumUsers(); i++ )
+		for ( int i = 0; i < long(GetNumUsers()); i++ )
 		{
 			User& user = GetUser( i );
 			UserBattleStatus& status = user.BattleStatus();
@@ -796,7 +796,7 @@ void Battle::Autobalance( BalanceType balance_type, bool support_clans, bool str
     if ( numallyteams == 0 ) // 0 == use num start rects
     {
         int ally = 0;
-        for ( int i = 0; i < GetNumRects(); ++i )
+        for ( int i = 0; i < long(GetNumRects()); ++i )
         {
             BattleStartRect sr = GetStartRect(i);
             if ( sr.IsOk() )
